JmeSurfaceView: migration to new package (com.jme3.view.surfaceview)
