+++ jme3-android/src/main/java/com/jme3/view/surfaceview/OnLayoutDrawn.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009-2020 jMonkeyEngine
+ * Copyright (c) 2009-2022 jMonkeyEngine
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,25 +29,23 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.jme3.app.jmeSurfaceView;
+package com.jme3.view.surfaceview;
 
+import android.view.View;
 import com.jme3.app.LegacyApplication;
-import com.jme3.system.AppSettings;
 
 /**
- * An interface used for invoking an event when the user delay finishes, on the first update of the game.
+ * An interface used for dispatching an event when the layout holding the {@link android.opengl.GLSurfaceView} is drawn,
+ * the event is dispatched on the user activity context thread.
  *
  * @author pavl_g.
- * @see JmeSurfaceView#setOnRendererCompleted(OnRendererCompleted)
  */
-public interface OnRendererCompleted {
+public interface OnLayoutDrawn {
     /**
-     * Invoked when the user delay finishes, on the first update of the game, the event is dispatched on the
-     * enclosing Activity context thread.
+     * Dispatched when the layout is drawn on the screen.
      *
-     * @param application the current jme game instance.
-     * @param appSettings the current window settings of the running jme game.
-     * @see JmeSurfaceView#update()
+     * @param legacyApplication the application instance.
+     * @param layout            the current layout.
      */
-    void onRenderCompletion(LegacyApplication application, AppSettings appSettings);
+    void onLayoutDrawn(LegacyApplication legacyApplication, View layout);
 }
